# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API",
        "description": "Products API",
        "termsOfService": "",
        "contact": {
            "email": "juanitotarroja@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of Products",
                "description": "Returns list of created Products",
                "operationId": "",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Target Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Rows per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Product name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[url_segment]",
                        "in": "query",
                        "description": "URL segment",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[sku]",
                        "in": "query",
                        "description": "Product SKU",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "current_page": 1,
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "FooBar DDR4 RAM 3200Mhz",
                                            "url_segment": "foobar-ddr4-ram-3200mhz",
                                            "sku": "32424234234",
                                            "price": 7200,
                                            "created_at": "2021-01-17 18:21:22",
                                            "updated_at": "2021-01-17 18:21:22"
                                        }
                                    ],
                                    "from": 1,
                                    "last_page": 4,
                                    "per_page": 1,
                                    "to": 1,
                                    "total": 4
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "errors": {
                                        "page": [
                                            "The page must be an integer."
                                        ],
                                        "per_page": [
                                            "The per page must be an integer."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create Products",
                "description": "Creates Products. Note that similar url_segments shall be suffixed by an integer at the end since url_segments MUST BE UNIQUE.",
                "operationId": "",
                "requestBody": {
                    "description": "ProductsPostRequest",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductsPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "name": "FooBar DDR4 RAM 3200Mhz",
                                        "url_segment": "foobar-ddr4-ram-3200mhz",
                                        "sku": "32424234234",
                                        "price": 7200,
                                        "created_at": "2021-01-17 18:21:22",
                                        "updated_at": "2021-01-17 18:21:22"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "The name must be a string."
                                        ],
                                        "url_segment": [
                                            "The url segment must be a string."
                                        ],
                                        "sku": [
                                            "The sku must be an integer."
                                        ],
                                        "price": [
                                            "The price must be a number."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/products/url_segment/{url_segment}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product by URL Segment",
                "description": "Get Product by URL Segment. Returns Product details",
                "operationId": "",
                "parameters": [
                    {
                        "name": "url_segment",
                        "in": "path",
                        "description": "Product URL Segment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "name": "FooBar DDR4 RAM 3200Mhz",
                                        "url_segment": "foobar-ddr4-ram-3200mhz",
                                        "sku": "32424234234",
                                        "price": 7200,
                                        "created_at": "2021-01-17 18:21:22",
                                        "updated_at": "2021-01-17 18:21:22"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "errors": {
                                        "url_segment": [
                                            "The url_segment must be a string."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductsPostRequest": {
                "title": "ProductsPostRequest",
                "description": "ProductsPostRequest",
                "properties": {
                    "name": {
                        "title": "product_name",
                        "description": "Product Name",
                        "type": "string",
                        "format": "string"
                    },
                    "url_segment": {
                        "title": "url_segment",
                        "description": "Product URL Segment",
                        "type": "string",
                        "format": "string"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "Product SKU",
                        "type": "string",
                        "format": "string"
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}